+++ src/core/system.hpp
@@ -98,8 +98,8 @@ namespace nanos
          ~System ();
 
          void submit ( WD &work );
-         void submitWithDependencies (WD& work, int numDeps, Dependency* deps);
-         void waitOn ( int numDeps, Dependency* deps);
+         void submitWithDependencies (WD& work, size_t numDeps, Dependency* deps);
+         void waitOn ( size_t numDeps, Dependency* deps);
          void inlineWork ( WD &work );
 
          void createWD (WD **uwd, size_t num_devices, nanos_device_t *devices,
